Start writing function to return guess feedback
Start to implement a function that returns 4 key pegs that reflect
the feedback of the codebreakers guess on the code.

This leaves a bug in which a white key peg is continually left for a
player's color guess, even if they have guessed more of the same color
than even exists in the code.

Implement a tally function that can be used to help resolve this
later on.